+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Added `with_tag_parents()` to provide parent information to tags.
 * `account()` can now load account information from a data.frame.
+* `account()` can now fetch accounts from a data.frame. 
+* Much debugging of how `filter_mentions()` handles multiple 
+  accounts, some with no brands.
 
 # brandseyer2 0.0.1.9009